requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
TTS>=0.20.0
pyyaml>=6.0
ffmpeg-python>=0.2.0
numpy>=1.24.0
soundfile>=0.12.0
requests>=2.31.0

